backend/connectors/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.1
boto3==1.35.94
reportlab==4.2.5
matplotlib==3.9.4
numpy==2.1.3
pypdf==5.1.0
httpx==0.28.1
cryptography>=42.0.0
